+++ app/src/app/components/Toolbar/DistrictrIcon.tsx
@@ -1,5 +1,5 @@
-import {SVGProps} from 'react';
-
+// TODO: Refine these icons / bring in more thoughtful visuals
+// this starts a component and config set up but is not complete
 const IconPaths = {
   lock: (
     <path d="m50 2.082c-14.699 0-26.82 10.77-28.832 25-0.023438 0-0.039063 0.003907-0.074219 0l-0.046875 0.78906c-0.125 1.1094-0.21094 2.2305-0.21094 3.375v4.168c-4.6016 0-8.332 3.75-8.332 8.332v45.832c0 4.582 3.7305 8.332 8.332 8.332h58.332c4.6055 0 8.332-3.75 8.332-8.332v-45.832c0-4.582-3.7305-8.332-8.332-8.332v-4.168c0-1.1484-0.085938-2.2656-0.21094-3.375l-0.046875-0.78906c-0.035156 0.003907-0.050781 0-0.074218 0-2.0156-14.23-14.133-25-28.832-25zm0 12.5c1.1484 0 2.2773 0.12109 3.3594 0.34375 5.4258 1.1172 9.8867 4.875 11.996 9.8555 0.41797 0.99219 0.74609 2.0312 0.96875 3.1094 0.22266 1.0859 0.34375 2.2109 0.34375 3.3594v4.168h-33.332v-4.168c0-1.1484 0.16797-2.3008 0.39062-3.3867 0.3125-1.5391 0.89063-2.9688 1.5938-4.3203 0.21484-0.40625 0.40625-0.82031 0.65234-1.2031 0.25391-0.40234 0.54688-0.76953 0.82812-1.1484 0.33984-0.44531 0.69141-0.88281 1.0742-1.293 0.26953-0.28906 0.55469-0.5625 0.84766-0.82813 0.47656-0.44141 0.97656-0.85937 1.5078-1.2461 0.28516-0.21094 0.57031-0.42188 0.87109-0.60938 0.58203-0.37109 1.1953-0.6875 1.8242-0.98438 0.29688-0.14062 0.58594-0.30078 0.89453-0.42188 0.67969-0.27344 1.3906-0.46875 2.1094-0.65234 0.30078-0.078125 0.58594-0.18359 0.89453-0.24219 1.0312-0.19922 2.0938-0.32422 3.1797-0.32422z" />
