+++ app/src/app/components/Toolbar/ExitBlockViewButtons.tsx
@@ -1,10 +1,8 @@
 import {useMapStore} from '@/app/store/mapStore';
 import {Button, Flex} from '@radix-ui/themes';
-
 export function ExitBlockViewButtons() {
   const captiveIds = useMapStore(store => store.captiveIds);
   const exitBlockView = useMapStore(store => store.exitBlockView);
-
   return captiveIds.size ? (
     <Flex mt={'1'} pt="1" className="border-black borer-t-2" direction={'column'} gap="2">
       <Button onClick={() => exitBlockView()} variant={'solid'}>
