+++ app/src/app/components/Toolbar/Toolbar.tsx
@@ -1,20 +1,19 @@
 'use client';
-import {Flex, IconButton} from '@radix-ui/themes';
-import * as Tooltip from '@radix-ui/react-tooltip';
+import {IconButton} from '@radix-ui/themes';
 import {useMapStore} from '@store/mapStore';
 import {MoveIcon, RotateCounterClockwiseIcon} from '@radix-ui/react-icons';
 import React, {useEffect, useLayoutEffect, useRef, useState} from 'react';
 import {ActiveTool} from '@constants/types';
 import Draggable from 'react-draggable';
 import {useToolbarStore} from '@/app/store/toolbarStore';
-import {ToolUtilities} from '@components/Toolbar/ToolUtilities';
-import {useActiveTools} from '@components/Toolbar/ToolbarUtils';
+import {ToolControls} from '@/app/components/Toolbar/ToolControls';
+import {useActiveTools} from '@/app/components/Toolbar/ToolUtils';
+import {ToolButtons} from './ToolButtons';
 
 export const Toolbar = () => {
   const activeTool = useMapStore(state => state.activeTool);
   const setActiveTool = useMapStore(state => state.setActiveTool);
   const [showShortcuts, setShowShortcuts] = useState(false);
-  const [activeTooltip, setActiveTooltip] = useState<ActiveTool | null>(null);
   const {
     x: userX,
     y: userY,
@@ -56,7 +55,7 @@ export const Toolbar = () => {
       containerRef.getBoundingClientRect().width / 2 - (toolbarWidth ?? 0) / 2,
       containerRef.getBoundingClientRect().height - 50
     );
-    setIsMobile(containerRef?.clientWidth < (activeTools.length * toolbarSize) * 2);
+    setIsMobile(containerRef?.clientWidth < activeTools.length * toolbarSize * 2);
   };
 
   useLayoutEffect(() => {
@@ -130,7 +129,7 @@ export const Toolbar = () => {
         onMouseEnter={() => setHovered(true)}
         onMouseLeave={() => setHovered(false)}
       >
-        <Tools
+        <ToolButtons
           showShortcuts={showShortcuts}
           isMobile={isMobile}
           toolbarItemsRef={toolbarItemsRef}
@@ -171,100 +170,8 @@ export const Toolbar = () => {
           </>
         )}
 
-        <ToolUtilities isMobile={isMobile} />
+        <ToolControls isMobile={isMobile} />
       </div>
     </Draggable>
-  );
-};
-
-const Tools: React.FC<{
-  showShortcuts: boolean;
-  toolbarItemsRef: React.RefObject<HTMLDivElement>;
-  isMobile?: boolean;
-}> = ({showShortcuts, isMobile, toolbarItemsRef}) => {
-  const activeTool = useMapStore(state => state.activeTool);
-  const setActiveTool = useMapStore(state => state.setActiveTool);
-  const [activeTooltip, setActiveTooltip] = useState<ActiveTool | null>(null);
-  const {rotation: userRotation, customizeToolbar, toolbarSize} = useToolbarStore(state => state);
-  const activeTools = useActiveTools();
-  const rotation = customizeToolbar && !isMobile ? userRotation : 'horizontal';
-
-  return (
-    <Flex
-      justify={'center'}
-      align="center"
-      ref={toolbarItemsRef}
-      direction={rotation === 'horizontal' ? 'row' : 'column'}
-      className="shadow-md overflow-hidden bg-white"
-      width="100%"
-    >
-      {activeTools.map((tool, i) => {
-        const IconComponent = tool.icon;
-        return (
-          <Tooltip.Provider key={`toolbar-tooltip-${i}`}>
-            <Tooltip.Root open={showShortcuts || activeTooltip === tool.mode || undefined}>
-              <Tooltip.Trigger
-                asChild
-                style={{
-                  flexGrow: isMobile ? 1 : undefined,
-                }}
-              >
-                <IconButton
-                  key={`${tool.mode}-flex`}
-                  className={`cursor-pointer ${i === 0 ? 'rounded-l-lg' : ''} ${
-                    i === activeTools.length - 1 ? 'rounded-r-lg' : ''
-                  }`}
-                  onMouseEnter={() => setActiveTooltip(tool.mode)}
-                  onMouseLeave={() => setActiveTooltip(null)}
-                  onClick={() => {
-                    if (tool.onClick) {
-                      tool.onClick();
-                    } else {
-                      setActiveTool(activeTool === tool.mode ? 'pan' : tool.mode);
-                    }
-                  }}
-                  style={{
-                    width: toolbarSize,
-                    height: toolbarSize,
-                    ...(tool?.iconStyle || {}),
-                  }}
-                  variant={tool.variant || activeTool === tool.mode ? 'solid' : 'surface'}
-                  color={tool.color}
-                  radius="none"
-                  disabled={tool.disabled}
-                >
-                  <IconComponent width={toolbarSize * 0.4} height={toolbarSize * 0.4} />
-                </IconButton>
-              </Tooltip.Trigger>
-              <Tooltip.Portal>
-                <Tooltip.Content
-                  side={rotation === 'horizontal' ? 'top' : 'right'}
-                  className="select-none rounded bg-gray-900 px-2 py-1 text-xs text-center text-white"
-                  sideOffset={5}
-                >
-                  {!showShortcuts && (
-                    <>
-                      {tool.label}
-                      <br />
-                    </>
-                  )}{' '}
-                  {rotation === 'horizontal' ? (
-                    tool.hotKeyLabel.split(' + ').map((key, i) => (
-                      <span key={i} className="text-xs">
-                        {key}
-                        <br />
-                      </span>
-                    ))
-                  ) : (
-                    <span className="text-xs">{tool.hotKeyLabel}</span>
-                  )}
-                  <Tooltip.Arrow className="fill-gray-900" />
-                </Tooltip.Content>
-              </Tooltip.Portal>
-            </Tooltip.Root>
-          </Tooltip.Provider>
-        );
-      })}
-    </Flex>
   );
 };
